// src/app/shared/interfaces.ts
export interface IWidget {
    id: number
    label: string
    content: Type<unknown>
    rows?: number
    columns?: number
}